+++ hass-addon-sunsynk/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## **2022.2.5-0.1.0** - 2022-02-05
+
+Major update
+
+Python sunsynk module 0.1.0:
+
+- Modified definitions
+- Writing registers
+
+Sunsynk Add-On:
+
+- Modify settings with presets (BETA)
+- New MQTT select entity
+- No more restart on Timeouts
+- Filter updates (support strings)
+
 ## **2022.1.26-0.0.8** - 2022-01-26
 
 - Added grid_voltage